Vector methods on Vecot
