+++ src/DSE.Open.Numerics/Sequence.PopulationVariance.cs
@@ -1,4 +1,4 @@
-﻿// Copyright (c) Down Syndrome Education International and Contributors. All Rights Reserved.
+// Copyright (c) Down Syndrome Education International and Contributors. All Rights Reserved.
 // Down Syndrome Education International and Contributors licence this file to you under the MIT license.
 
 using System.Numerics;
@@ -13,7 +13,7 @@ public static partial class Sequence
     {
         if (sequence.TryGetSpan(out var span))
         {
-            return VectorPrimitives.PopulationVariance(span, mean);
+            return Vector.PopulationVariance(span, mean);
         }
 
         throw new NotImplementedException();
