+++ src/DSE.Open.Numerics/Vector.Quantiles.cs
@@ -5,24 +5,19 @@ using System.Numerics;
 
 namespace DSE.Open.Numerics;
 
-public static partial class VectorPrimitives
+public partial class Vector
 {
     /// <summary>
-    /// Gets the sample variance of a sequence of at least two numbers.
+    /// Divide data into n continuous intervals with equal probability. Returns a list of n - 1
+    /// cut points separating the intervals.
     /// </summary>
     /// <typeparam name="T">The type of the elements in the sequence.</typeparam>
     /// <param name="span">The sequence of elements to use for the calculation.</param>
-    /// <param name="mean"></param>
+    /// <param name="n"></param>
     /// <returns></returns>
-    public static T PopulationVariance<T>(ReadOnlySpan<T> span, T? mean = default)
+    /// <exception cref="NotImplementedException"></exception>
+    public static ReadOnlySpan<T> Quantiles<T>(ReadOnlySpan<T> span, int n = 4)
         where T : struct, INumberBase<T>
-    {
-        return PopulationVariance<T, T>(span, mean);
-    }
-
-    public static TResult PopulationVariance<T, TResult>(ReadOnlySpan<T> span, T? mean = default)
-        where T : struct, INumberBase<T>
-        where TResult : struct, INumberBase<TResult>
     {
         throw new NotImplementedException();
     }
