+++ src/DSE.Open.Numerics/Vector.Variance.cs
@@ -5,7 +5,7 @@ using System.Numerics;
 
 namespace DSE.Open.Numerics;
 
-public static partial class VectorPrimitives
+public partial class Vector
 {
     /// <summary>
     /// Gets the sample variance of a sequence of at least two numbers.
