+++ test/DSE.Open.Numerics.Tests/VectorTests.Mean.cs
@@ -3,13 +3,13 @@
 
 namespace DSE.Open.Numerics;
 
-public partial class VectorPrimitivesTests
+public partial class VectorTests
 {
     [Fact]
     public void Mean_Span_Int32_Int32()
     {
         ReadOnlySpan<int> sequence = [1, 2, 3, 4, 5, 6, 7, 8, 9, 10];
-        var mean = VectorPrimitives.Mean<int, int>(sequence);
+        var mean = Vector.Mean<int, int>(sequence);
         Assert.Equal(5, mean);
     }
 
@@ -17,7 +17,7 @@ public partial class VectorPrimitivesTests
     public void Mean_Span_Int32_Double()
     {
         ReadOnlySpan<int> sequence = [1, 2, 3, 4, 5, 6, 7, 8, 9, 10];
-        var mean = VectorPrimitives.Mean<int, double>(sequence);
+        var mean = Vector.Mean<int, double>(sequence);
         Assert.Equal(5.5, mean);
     }
 }
